Unreachable Code issue fixed and gameCost data type changed to decimal

+++ SignalR/SignalR.Server/GameRoom.cs
@@ -4,9 +4,9 @@
     {
         public string RoomName { get; set; }
         public string GameType { get; set; }
-        public int GameCost { get; set; }
+        public decimal GameCost { get; set; }
         public List<User> Users { get; set; }
-        public GameRoom(string roomName, string gameType, int gameCost)
+        public GameRoom(string roomName, string gameType, decimal gameCost)
         {
             RoomName = roomName;
             GameType = gameType;
